mainly a function calculate the init value of gloV

+++ IR/include/ConstantFloat.h
@@ -6,14 +6,14 @@ class Context;
 
 class ConstantFloat : public Constant {
 public:
-  ConstantFloat(Context &c, Type *type, double value)
+  ConstantFloat(Context &c, Type *type, float value)
       : Constant(c, type, "", 0), value_(value) {}
   double get_value() { return value_; }
   std::string print() override;
 
 public:
-  static ConstantFloat *get(Context &context, double value);
+  static ConstantFloat *get(Context &context, float value);
 
 private:
-  double value_;
+  float value_;
 };
